indexer/requirements.txt
asyncpg
meilisearch
python-dotenv
